+++ config.json
@@ -3,5 +3,5 @@
     "voiceChannelId": "1275842522177863734",
     "stream": "https://stream.zeno.fm/t4is3umfojdtv.mp3",
     "prefix": "kf!",
-    "volume": 25
+    "volume": 100
 }
