+++ package.json
@@ -9,7 +9,9 @@
     "dotenv": "^16.4.5",
     "ffmpeg-static": "^5.2.0",
     "sodium-native": "^4.1.1",
-    "sqlite3": "^5.1.7"
+    "sqlite": "^5.1.1",
+    "sqlite3": "^5.1.7",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.1.0",
