+++ src/index.ts
@@ -1,4 +1,5 @@
 import { RadioClient } from './utils/client';
 import 'dotenv/config';
+
 const client = new RadioClient();
 client.login(process.env.TOKEN);
